refactor file structure
add model dir for data classes and subdir filebased for frame group/chunk dataclasses

move detection data classes and frame data classes to own files
