+++ tests/track/test_frame_group.py
@@ -4,7 +4,7 @@ from typing import Any
 from unittest.mock import MagicMock, patch
 
 from OTVision.dataformat import EXPECTED_DURATION, FILENAME, RECORDED_START_DATE, VIDEO
-from OTVision.track.frame_group import FrameGroup
+from OTVision.track.model.filebased.frame_group import FrameGroup
 from tests.track.test_data_builder import (
     DEFAULT_HOSTNAME,
     DEFAULT_INPUT_FILE_PATH,
